move rendering retry into image service
